STSModAutoTest/requirements.txt
airtest
Pillow
